restructure render_status.py
