+++ flow/project.py
@@ -68,7 +68,7 @@ from .labels import classlabel
 from .labels import _is_label_func
 from . import legacy
 from .util import config as flow_config
-
+from . import render_status
 
 logger = logging.getLogger(__name__)
 
@@ -946,7 +946,7 @@ class FlowProject(signac.contrib.Project, metaclass=_FlowProjectClass):
                      unroll=True, compact=False, pretty=False,
                      file=None, err=None, ignore_errors=False,
                      no_parallelize=False, template=None, profile=False,
-                     eligible_jobs_max_lines=None):
+                     eligible_jobs_max_lines=None, option='terminal'):
         """Print the status of the project.
 
         .. versionchanged:: 0.6
@@ -1036,132 +1036,12 @@ class FlowProject(signac.contrib.Project, metaclass=_FlowProjectClass):
         if jobs is None:
             jobs = self     # all jobs
 
-        # use Jinja2 template for status output
-        if template is None:
-            if detailed and expand:
-                template = 'status_expand.jinja'
-            elif detailed and not unroll:
-                template = 'status_stack.jinja'
-            elif detailed and compact:
-                template = 'status_compact.jinja'
-            else:
-                template = 'status.jinja'
-
         if eligible_jobs_max_lines is None:
             eligible_jobs_max_lines = flow_config.get_config_value('eligible_jobs_max_lines')
 
         # initialize jinja2 template evnronment and necessary filters
         template_environment = self._template_environment()
 
-        def draw_progressbar(value, total, width=40):
-            """Visualize progess with a progress bar.
-
-            :param value:
-                The current progress as a fraction of total.
-            :type value:
-                int
-            :param total:
-                The maximum value that 'value' may obtain.
-            :type total:
-                int
-            :param width:
-                The character width of the drawn progress bar.
-            :type width:
-                int
-            """
-
-            assert value >= 0 and total > 0
-            ratio = ' %0.2f%%' % (100 * value / total)
-            n = int(value / total * width)
-            return '[' + ''.join(['#'] * n) + ''.join(['-'] * (width - n)) + ']' + ratio
-
-        def job_filter(job_op, scheduler_status_code, all_ops):
-            """filter eligible jobs for status print.
-
-            :param job_ops:
-                Operations information for a job.
-            :type job_ops:
-                OrderedDict
-            :param scheduler_status_code:
-                Dictionary information for status code
-            :type scheduler_status_code:
-                Dictionary
-            :param all_ops:
-                Boolean value indicate if all operations should be displayed
-            :type all_ops:
-                Boolean
-            """
-
-            if scheduler_status_code[job_op['scheduler_status']] != 'U' or \
-               job_op['eligible'] or all_ops:
-                return True
-            else:
-                return False
-
-        def get_operation_status(operation_info, symbols):
-            """Determine the status of an operation.
-
-            :param operation_info:
-                Dicionary containing operation information
-            :type operation_info:
-                Dictionary
-            :param symbols:
-                Dicionary containing code for different job status
-            :type symbols:
-                Dictionary
-            """
-
-            if operation_info['scheduler_status'] >= JobStatus.active:
-                op_status = u'running'
-            elif operation_info['scheduler_status'] > JobStatus.inactive:
-                op_status = u'active'
-            elif operation_info['completed']:
-                op_status = u'completed'
-            elif operation_info['eligible']:
-                op_status = u'eligible'
-            else:
-                op_status = u'ineligible'
-
-            return symbols[op_status]
-
-        if pretty:
-            def highlight(s, eligible):
-                """Change font to bold within jinja2 template
-
-                :param s:
-                    The string to be printed
-                :type s:
-                    str
-                :param eligible:
-                    Boolean value for job eligibility
-                :type eligible:
-                    Boolean
-                """
-                if eligible:
-                    return '**' + s + '**'
-                else:
-                    return s
-        else:
-            def highlight(s, eligible):
-                """Change font to bold within jinja2 template
-
-                :param s:
-                    The string to be printed
-                :type s:
-                    str
-                :param eligible:
-                    Boolean value for job eligibility
-                :type eligible:
-                    boolean
-                """
-                return s
-
-        template_environment.filters['highlight'] = highlight
-        template_environment.filters['draw_progressbar'] = draw_progressbar
-        template_environment.filters['get_operation_status'] = get_operation_status
-        template_environment.filters['job_filter'] = job_filter
-
-        template = template_environment.get_template(template)
         context = self._get_standard_template_context()
 
         # get job status information
@@ -1393,17 +1273,15 @@ class FlowProject(signac.contrib.Project, metaclass=_FlowProjectClass):
         if n > 0:
             context['op_counter'].append(('[{} more operations omitted]'.format(n), ''))
 
-        markdown_output = template.render(**context)
-        # print(markdown_output, file=file)
-        # add logic for html output here
-        # html_output = mistune.markdown(markdown_output)
-        # print(html_output, file=file)
-        terminal_output = mdv.main(markdown_output)
-        print(terminal_output, file=file)
+        render_result = render_status._render_status()
+        render_output = render_result.render(template, template_environment, context, file, detailed, expand, unroll, compact, pretty, option)
+
 
         # Show profiling results (if enabled)
         if profiling_results:
             print('\n' + '\n'.join(profiling_results), file=file)
+
+        return render_output
 
     def run_operations(self, operations=None, pretend=False, np=None, timeout=None, progress=False):
         """Execute the next operations as specified by the project's workflow.
@@ -2182,6 +2060,11 @@ class FlowProject(signac.contrib.Project, metaclass=_FlowProjectClass):
             '--no-parallelize',
             action='store_true',
             help="Do not parallelize the status determination.")
+        view_group.add_argument(
+            '-op', '--option',
+            type=str,
+            default='terminal',
+            help="define rendering format.")
 
     def labels(self, job):
         """Yields all labels for the given ``job``.
